[#71] Rename : Color, Theme, Type 을 designsystem 모듈로 이동
