+++ core/designsystem/src/main/kotlin/com/dhc/designsystem/Theme.kt
@@ -1,4 +1,4 @@
-package com.dhc.dhcandroid.ui.theme
+package com.dhc.designsystem
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
