+++ core/designsystem/src/main/kotlin/com/dhc/designsystem/Type.kt
@@ -1,4 +1,4 @@
-package com.dhc.dhcandroid.ui.theme
+package com.dhc.designsystem
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
